move clone into its own file
