+++ src/process-events.c
@@ -13,10 +13,7 @@
 #include "common.h"
 #include "process/unshare.h"
 #include "process/helpers.h"
-
-#define CLONE_ARGS_SIZE_VER0 64 /* sizeof first published struct */
-#define CLONE_ARGS_SIZE_VER1 80 /* sizeof second published struct */
-#define CLONE_ARGS_SIZE_VER2 88 /* sizeof third published struct */
+#include "process/clone.h"
 
 // maximum size of each buffer (since they may have flexible arrays at
 // the end)
@@ -26,14 +23,6 @@
 
 // maximum segments we can read from a d_path before doing a tail call
 #define MAX_PATH_SEGMENTS_NOTAIL 32
-
-// these structs are used to send data gathered/calculated in a kprobe
-// to the kretprobe.
-
-typedef struct {
-    u64 pid_tgid;
-    clone_info_t clone_info;
-} incomplete_clone_t;
 
 // used for events with flexible sizes (i.e., exec*) so it can send
 // extra data. Used in conjuction with a map such that it does not use
@@ -66,17 +55,6 @@ struct bpf_map_def SEC("maps/percpu_path") percpu_path = {
     .key_size = sizeof(u32),
     .value_size = sizeof(cached_path_t),
     .max_entries = 1,
-    .pinning = 0,
-    .namespace = "",
-};
-
-// A map of clones that have started (a kprobe) but are yet to finish
-// (the kretprobe).
-struct bpf_map_def SEC("maps/incomplete_clone") incomplete_clones = {
-    .type = BPF_MAP_TYPE_LRU_HASH,
-    .key_size = sizeof(u64),
-    .value_size = sizeof(incomplete_clone_t),
-    .max_entries = 8 * 1024,
     .pinning = 0,
     .namespace = "",
 };
@@ -557,77 +535,6 @@ int kretprobe__ret_sys_execveat_4_8(struct pt_regs *ctx)
     return 0;
 }
 
-static __always_inline void enter_clone(struct pt_regs *ctx, process_message_type_t pm_type, unsigned long flags)
-{
-    // we do not care about threads spawning; ignore clones that would
-    // share the same thread group as the parent.
-    if (flags & CLONE_THREAD)
-        return;
-
-    u64 pid_tgid = bpf_get_current_pid_tgid();
-    incomplete_clone_t event = {0};
-    event.pid_tgid = pid_tgid;
-    event.clone_info.flags = flags;
-
-    int ret = bpf_map_update_elem(&incomplete_clones, &pid_tgid, &event, BPF_ANY);
-    if (ret < 0)
-    {
-        process_message_t pm = {0};
-        pm.type = PM_WARNING;
-        pm.u.warning_info.pid_tgid = pid_tgid;
-        pm.u.warning_info.message_type = pm_type;
-        pm.u.warning_info.code = PMW_UPDATE_MAP_ERROR;
-        pm.u.warning_info.info.err = ret;
-
-        push_message(ctx, &pm);
-
-        return;
-    }
-
-    return;
-}
-
-// handles the kretprobe of clone-like syscalls (fork, vfork, clone, clone3)
-static __always_inline void exit_clonex(struct pt_regs *ctx, pprocess_message_t pm, process_message_type_t pm_type)
-{
-    u64 pid_tgid = bpf_get_current_pid_tgid();
-    load_event(incomplete_clones, pid_tgid, incomplete_clone_t);
-
-    int retcode = PT_REGS_RC(ctx);
-    if (retcode < 0)
-        goto Done;
-
-    void *ts = (void *)bpf_get_current_task();
-    int ret = fill_syscall(&pm->u.syscall_info, ts, pid_tgid >> 32);
-    if (ret > 0) return;
-    if (ret < 0) goto EmitWarning;
-
-    pm->type = pm_type;
-    pm->u.syscall_info.data.clone_info = event.clone_info;
-    pm->u.syscall_info.retcode = retcode;
-
-    push_message(ctx, pm);
-
-    goto Done;
-
- EventMismatch:;
-    error_info_t info = {0};
-    info.stored_pid_tgid = event.pid_tgid;
-    set_local_warning(PMW_PID_TGID_MISMATCH, info);
-
- EmitWarning:;
-    push_warning(ctx, pm, pm_type);
-
- Done:;
-    // only delete at the every end so the event pointer above is
-    // valid for the duration of this function.
-    bpf_map_delete_elem(&incomplete_clones, &pid_tgid);
-    return;
-
- NoEvent:;
-    return;
-}
-
 SEC("kprobe/sys_clone_4_8")
 #if defined(__TARGET_ARCH_x86)
 int BPF_KPROBE_SYSCALL(kprobe__sys_clone_4_8, unsigned long flags, void __user *stack,
@@ -657,7 +564,7 @@ SEC("kretprobe/ret_sys_clone3")
 int kretprobe__ret_sys_clone3(struct pt_regs *ctx)
 {
     process_message_t sev = {0};
-    exit_clonex(ctx, &sev, PM_CLONE3);
+    exit_clone(ctx, &sev, PM_CLONE3);
     return 0;
 }
 
@@ -669,48 +576,10 @@ int kretprobe__ret_sys_clone3(struct pt_regs *ctx)
 SEC("kprobe/read_pid_task_struct")
 int kprobe__read_pid_task_struct(struct pt_regs *ctx)
 {
-    u64 pid_tgid = bpf_get_current_pid_tgid();
-    load_event(incomplete_clones, pid_tgid, incomplete_clone_t);
-
     // get passed in task_struct
     void *ts = (void *)PT_REGS_PARM1(ctx);
+    add_childpid(ts);
 
-    // get the true pid
-    u32 npid = 0;
-    if (read_field(ts, CRC_TASK_STRUCT_PID, &npid, sizeof(npid)) < 0) {
-        // don't bother emitting a warning; worst case we can use the retcode
-        return 0;
-    }
-
-    u32 ntgid = 0;
-    if (read_field(ts, CRC_TASK_STRUCT_TGID, &ntgid, sizeof(ntgid)) < 0) {
-        // don't bother emitting a warning; worst case we can use the retcode
-        return 0;
-    }
-
-    // this means that this task_struct belongs to a non-main
-    // thread. We do not care about new threads being spawned so exit
-    // early.
-    if (npid != ntgid)
-    {
-        // the kretprobe shouldn't care about it either
-        bpf_map_delete_elem(&incomplete_clones, &pid_tgid);
-        return 0;
-    }
-
-    // deliberately not deleting from the map - we'll let the
-    // kretprobe do that and send the event
-    event.clone_info.child_pid = ntgid;
-    // we copied the event so we need to manually update it
-    bpf_map_update_elem(&incomplete_clones, &pid_tgid, &event, BPF_ANY);
-
-    return 0;
-
- EventMismatch:;
-    // let the kretprobe return the error as that has more information
-    return 0;
-
- NoEvent:;
     return 0;
 }
 
@@ -734,7 +603,7 @@ SEC("kretprobe/ret_sys_clone")
 int kretprobe__ret_sys_clone(struct pt_regs *ctx)
 {
     process_message_t sev = {0};
-    exit_clonex(ctx, &sev, PM_CLONE);
+    exit_clone(ctx, &sev, PM_CLONE);
     return 0;
 }
 
@@ -742,7 +611,7 @@ SEC("kretprobe/ret_sys_fork")
 int kretprobe__ret_sys_fork(struct pt_regs *ctx)
 {
     process_message_t sev = {0};
-    exit_clonex(ctx, &sev, PM_FORK);
+    exit_clone(ctx, &sev, PM_FORK);
     return 0;
 }
 
@@ -750,7 +619,7 @@ SEC("kretprobe/ret_sys_vfork")
 int kretprobe__ret_sys_vfork(struct pt_regs *ctx)
 {
     process_message_t sev = {0};
-    exit_clonex(ctx, &sev, PM_VFORK);
+    exit_clone(ctx, &sev, PM_VFORK);
     return 0;
 }
